+++ src/emails/tasks/find_email.rs
@@ -1,6 +1,6 @@
-// find_email.rs - v10
+// find_email.rs - v11
 
-fn fold_find_email_10_0(x:&str)->Result<String>{Ok(x.to_string())}
-fn fold_find_email_10_0_check(y:&[u8])->bool{!y.is_empty()}
-struct FIND_EMAIL_10Inner0{val:u64,name:String}
-impl FIND_EMAIL_10Inner0{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}
+fn get_find_email_11_0(x:&str)->Result<String>{Ok(x.to_string())}
+fn get_find_email_11_0_check(y:&[u8])->bool{!y.is_empty()}
+struct FIND_EMAIL_11Inner0{val:u64,name:String}
+impl FIND_EMAIL_11Inner0{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}