+++ CHANGELOG.md
@@ -183,3 +183,4 @@
 - 2023-11-24: Init conf, bettercontact, follow up, qualify lead, +3
 - 2023-11-24: Docs configuration, conf, contacts, leads, +2
 - 2023-11-25: Shutdown mod, discovery, deal, discover, +5
+- 2023-11-26: Bench find email, discovery, follow up, contacts, +5